refactor: lift content up to avoid rerender
see: https://overreacted.io/before-you-memo/
