+++ src/page/Editor/components/ActionEditor/interface.tsx
@@ -1,4 +1,10 @@
-import { HTMLAttributes } from "react"
+import { HTMLAttributes, ReactNode } from "react"
+
 export interface ActionEditorProps extends HTMLAttributes<HTMLDivElement> {
   className?: string
+}
+
+export interface ActionEditorLayoutProps extends HTMLAttributes<HTMLDivElement> {
+  queryList: ReactNode,
+  actionEditorPanel: ReactNode
 }
